Make UpdateBackwardFunctor::B_ constant
Since each thread has its own copy of B_, it is normally possible
to update B_[i] without having to copy it into another Matrix.
But for unknown reasons, this does not work. Everything works
fine when adding some debug statement to print B_ content, but
computations are wrong when B_ is not displayed.
So make B_ constant and use a copy.